requirements.txt
pyspark
mlflow
scikit-learn
pandas
apache-airflow==2.8.1
duckdb
